requirements/base.txt
distribute==0.6.45
dj-database-url==0.2.1
django-crispy-forms==1.3.1
django-floppyforms==1.1
django-tables2==0.13.0
Django==1.5.1
psycopg2==2.5
python-dateutil==1.5
South==0.8.1
twilio==3.5.1
Unipath==1.0
Sphinx==1.2b1
gunicorn==0.17.4
